+++ ovp_uploads/serializers.py
@@ -30,7 +30,7 @@ class UploadedImageSerializer(serializers.ModelSerializer):
     model = UploadedImage
     fields = ('id', 'user', 'image', 'image_url', 'image_small_url', 'image_medium_url', 'image_large_url')
     read_only_fields = ('image_small', 'image_medium', 'image_large')
-    extra_kwargs = {'image': {'write_only': True}}
+    extra_kwargs = {'image': {'write_only': True}, 'crop_rect': {'write_only': True}}
 
   def get_image_url(self, obj):
     return build_absolute_uri(self.context['request'], obj.image)