+++ ovp_uploads/views.py
@@ -1,3 +1,5 @@
+import json
+
 from ovp_uploads.models import UploadedImage
 from ovp_uploads.serializers import UploadedImageSerializer, ImageGallerySerializer
 
@@ -5,6 +7,9 @@ from rest_framework import mixins
 from rest_framework import viewsets
 from rest_framework import response
 from rest_framework import status
+
+from .helpers import perform_image_crop
+
 
 class UploadedImageViewSet(mixins.CreateModelMixin, viewsets.GenericViewSet):
   queryset = UploadedImage.objects.all()
@@ -18,6 +23,13 @@ class UploadedImageViewSet(mixins.CreateModelMixin, viewsets.GenericViewSet):
 
     upload_header = request.META.get('HTTP_X_UNAUTHENTICATED_UPLOAD', None)
     is_authenticated = request.user.is_authenticated()
+
+    if request.data.get('crop_rect', None):
+      crop_rect = request.data.get('crop_rect')
+      if isinstance(crop_rect, str):
+        crop_rect = json.loads(crop_rect)
+      upload_data['image'] = perform_image_crop(upload_data['image'], crop_rect)
+      request.FILES['image'] = upload_data['image']
 
     if is_authenticated or upload_header:
       if upload_header: